Prevent duplicates when you watch both page and subpage.
This uses the DISTINCT option. MediaWiki core has code that creates
a partial group by in such a case. Manually override $wgSQLMode to
make everything work (especially unit tests) until core is fixed.

This also attempts to add a range in the hope that mysql can then
use dynamic range optimization when appropriate. It didn't seem to
work locally, but i think that's because mysql will only do it on
really big tables.

Change-Id: I8e8c8f81bace603ba26a1c348843c689efcf3f0e